QuizComponent - don't increment question until all subquestions answered
Previously, the quiz component would have incremented the active
question card whenever an answer was selected, even if that answer was
part of a subquestion. This change adds an additional prop, `onComplete`
that will only be called when all the subquestions have been answered,
or else on the top level first answer.