Refactor - introduce ScoreCalculation abstraction to track score
I was having some trouble extending the existing components to track
scores for questions nested under answers. This abstraction will provide
a stable base to make things more complicated.

We could've told that something was wrong with the existing abstractions
because:

1. `props.initialQuiz` - this is weird, the quiz itself isn't changing
2. passing two functions as props, `onAnswer` and `onComplete` and being
confused about when they would be called, what their arguments were,
etc.

The `ScoreCalculation` object lets us pass down both those functions but
in a self-desrcriptive manner, so that we can easily look up the
functions being called. It also provides an easy way to register
multiple callbacks, and removes some of the responsibility for knowing
when a question is completed.