Questions - add different sub questions modes
For some questions, all the sub questions should be displayed at once
and the user answers only one of them. This change makes the `Question`
class aware of the type to determine if it's been completed or not, and
renders a different sub question component if we need to display them
all at once.