ScoreCalculation - pull in isCompleted and computeScore logic
There was kind of a circular dependency between `ScoreCalculation` and
`Question` where `ScoreCalculation` would pass some of its internal data
to `Question`. That leads to a situation where if we change
`ScoreCalculation` we're likely to have to change `Question`. The data
structure is about to get more complicated to accommodate multiple
choice questions, so I want to keep more of that knowledge internal to
the calculation.