ScoreCalculation - extract class for tracking entries
Score calculation was getting difficult to test, because the only window
we had into it was the callbacks. It also mixed responsibilities - it
maintained data about answers, and it maintained a set of callbacks.
Now, it delegates the data to an `EntrySet`, which can hold some more
complexity for multiple choice questions.