SelectedAnswers - fix computeScore for sub-subquestions of answers
The kids question has sub questions of subquestions - "where will they
go to school" is broken down into different parts. The score was being
computed as zero for that question because we didn't recurse if
`question.hasSubQuestions()` for subquesitons of answers. That recursion
only happened at the top level, in `EntrySet`.