ScoreCalculation - fix determining completed multiple choice subs
There was a problem checking if a question was completed when it had an
answer with multiple choice subquestions. It determined completion too
early, because for answers we just iterated over the subquestions to
see if they were answered, not completed. That's a holdover from when we
had two classes for holding the selected answers, because one class only
looked at `isAnswered`.